+++ PCLParaphernalia/PrnParsePCLBinary.cs
@@ -3608,7 +3608,7 @@ namespace PCLParaphernalia
                     uint16Val = (ushort)((buf[offset] * 256) + buf[offset + 1]);
 
                     codeDesc = uint16Val.ToString() + " (-> bits per pixel = " +
-                               Math.Ceiling((Math.Log((double)uint16Val, 2))).ToString() +
+                               Math.Ceiling((Math.Log(uint16Val, 2))).ToString() +
                                ")";
 
                     PrnParseCommon.addTextRow(
