+++ PCLParaphernalia/PrnParsePCLXL.cs
@@ -4284,7 +4284,7 @@ namespace PCLParaphernalia
                             //                                                //
                             //------------------------------------------------//
 
-                            _attrDataVal = (int)iTot;
+                            _attrDataVal = iTot;
                         }
                     }
                 }
