+++ PCLParaphernalia/ToolPrintArea.xaml.cs
@@ -669,7 +669,7 @@ namespace PCLParaphernalia
         {
             pdlOptionsStore();
 
-            ToolPrintAreaPersist.saveDataCommon((int)_indxPDL);
+            ToolPrintAreaPersist.saveDataCommon(_indxPDL);
 
             ToolPrintAreaPersist.saveDataPCL("PCL",
                                              _indxOrientationPCL,
@@ -1170,7 +1170,7 @@ namespace PCLParaphernalia
                 lengthPrintable = (ushort)(sizeLongEdge -
                                           (marginsUnprintable * 2));
 
-                lengthLogical = (ushort)(sizeLongEdge);
+                lengthLogical = sizeLongEdge;
             }
             else
             {
@@ -1185,7 +1185,7 @@ namespace PCLParaphernalia
                 lengthPrintable = (ushort)(sizeShortEdge -
                                           (marginsUnprintable * 2));
 
-                lengthLogical = (ushort)(sizeShortEdge);
+                lengthLogical = sizeShortEdge;
             }
 
             //----------------------------------------------------------------//
