+++ PCLParaphernalia/ToolSoftFontGenPCLCommon.cs
@@ -534,7 +534,7 @@ namespace PCLParaphernalia
 
             flagOK = writeHddrSegHddr(pdlIsPCLXL,
                                        fmt16,
-                                       (uint)cSizeSegCC,
+                                       cSizeSegCC,
                                        segId,
                                        ref sumMod256);
 
@@ -621,7 +621,7 @@ namespace PCLParaphernalia
 
             flagOK = writeHddrSegHddr(pdlIsPCLXL,
                                        fmt16,
-                                       (uint)cSizeSegGC,
+                                       cSizeSegGC,
                                        segId,
                                        ref sumMod256);
 
@@ -687,7 +687,7 @@ namespace PCLParaphernalia
 
             flagOK = writeHddrSegHddr(pdlIsPCLXL,
                                        fmt16,
-                                       (uint)segLenGT,
+                                       segLenGT,
                                        segId,
                                        ref sumMod256);
 
@@ -714,7 +714,7 @@ namespace PCLParaphernalia
                                           (ushort)numTables,
                                           ref sumMod256);
 
-                crntOffset = (uint)(cSizeSegGTDirHddr + sizeDirectory);
+                crntOffset = cSizeSegGTDirHddr + sizeDirectory;
 
                 //------------------------------------------------------------//
                 //                                                            //
@@ -1288,7 +1288,7 @@ namespace PCLParaphernalia
 
             flagOK = writeHddrSegHddr(pdlIsPCLXL,
                                        fmt16,
-                                       (uint)cSizeSegPA,
+                                       cSizeSegPA,
                                        segId,
                                        ref sumMod256);
 
@@ -1370,7 +1370,7 @@ namespace PCLParaphernalia
 
             flagOK = writeHddrSegHddr(pdlIsPCLXL,
                                        fmt16,
-                                       (uint)cSizeSegVR,
+                                       cSizeSegVR,
                                        segId,
                                        ref sumMod256);
